pyproject.toml
[project]
name = "fyp2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=23.2.0",
    "alembic>=1.15.1",
    "fastapi>=0.119.0",
    "flwr-datasets[vision]>=0.5.0",
    "ipykernel>=7.0.1",
    "langchain>=1.0.0",
    "langchain-community>=0.3.31",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=3.1.0",
    "langchain-huggingface>=0.3.1",
    "langchain-postgres>=0.0.15",
    "langgraph>=0.6.10",
    "matplotlib>=3.10.7",
    "pandas>=2.3.2",
    "pgvector>=0.3.6",
    "plotly>=6.3.1",
    "pypdfium2>=4.30.0",
    "pytest>=8.4.2",
    "pytorch-lightning>=2.5.5",
    "pyyaml>=6.0.2",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "tensorboard>=2.20.0",
    "tensorboardx>=2.6.4",
    "torch",
    "torchvision",
    "torchaudio",
    "uvicorn>=0.37.0",
    "wandb>=0.22.2",
    "flwr[simulation]>=1.22.0",
    "flwr-datasets[vision]>=0.5.0",
    "python-multipart>=0.0.20",
    "psycopg2>=2.9.11",
    "tomli-w>=1.2.0",
    "tomli>=2.3.0",
    "sentence-transformers>=5.1.2",
    "langchain-mcp-adapters>=0.2.1",
    "reportlab>=4.4.9",
    "pre-commit>=4.5.1",
    "pytest-asyncio>=1.3.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["federated_pneumonia_detection*"]
exclude = ["data*", "wandb*", "docker*", "alembic*", "results*", "analysis*", "xray-vision-ai-forge*"]

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }
torchaudio = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "wandb",
    "data",
    "docker",
    "results",
    "analysis",
    "xray-vision-ai-forge",
]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.ruff.lint.per-file-ignores]
# Lines where adding # noqa: E501 makes them even longer than 88 chars
# These are docstrings/comments that should ideally be refactored but are acceptable
"federated_pneumonia_detection/src/api/endpoints/experiments/utils/__init__.py" = ["E501"]
"federated_pneumonia_detection/src/api/endpoints/inference/batch_prediction_endpoints.py" = ["E501"]
"federated_pneumonia_detection/src/api/endpoints/schema/runs_schemas.py" = ["E501"]
"federated_pneumonia_detection/src/api/endpoints/streaming/websocket_server.py" = ["E501"]
"federated_pneumonia_detection/src/api/middleware/error_handler.py" = ["E501"]
"federated_pneumonia_detection/src/boundary/CRUD/run.py" = ["E501"]
"federated_pneumonia_detection/src/boundary/CRUD/run_metric.py" = ["E501"]
"federated_pneumonia_detection/src/boundary/CRUD/server_evaluation.py" = ["E501"]
"federated_pneumonia_detection/src/boundary/services/cleanup_database.py" = ["E501"]
"federated_pneumonia_detection/src/boundary/cleanup_database.py" = ["E501"]
"federated_pneumonia_detection/src/boundary/engine.py" = ["E501"]
"federated_pneumonia_detection/src/control/agentic_systems/multi_agent_systems/chat/research_prompts.py" = ["E501"]
"federated_pneumonia_detection/src/control/agentic_systems/multi_agent_systems/chat/prompts/research_prompts.py" = ["E501"]
"federated_pneumonia_detection/src/control/agentic_systems/multi_agent_systems/chat/utils/titles.py" = ["E501"]
"federated_pneumonia_detection/src/control/agentic_systems/multi_agent_systems/chat/providers/titles.py" = ["E501"]
"federated_pneumonia_detection/src/control/analytics/facade.py" = ["E501"]
"federated_pneumonia_detection/src/control/analytics/internals/services/final_epoch_stats_service.py" = ["E501"]
"federated_pneumonia_detection/src/control/analytics/internals/utils/transformers.py" = ["E501"]
"federated_pneumonia_detection/src/control/dl_model/centralized_trainer.py" = ["E501"]
"federated_pneumonia_detection/src/control/dl_model/centralized_trainer_utils/db_operations.py" = ["E501"]
"federated_pneumonia_detection/src/control/dl_model/internals/model/callbacks/early_stopping.py" = ["E501"]
"federated_pneumonia_detection/src/control/dl_model/internals/model/callbacks/setup.py" = ["E501"]
"federated_pneumonia_detection/src/control/dl_model/internals/model/training_callbacks.py" = ["E501"]
"federated_pneumonia_detection/src/control/dl_model/internals/model/utils/metrics_handler.py" = ["E501"]
"federated_pneumonia_detection/src/control/federated_new_version/core/custom_strategy.py" = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["federated_pneumonia_detection"]
